Fix - missing imports
Missing imports (worked in browser because available on window global object)
