+++ packages/oo7-substrate/src/types.js
@@ -1,3 +1,20 @@
+function toLE(val, bytes) {
+	let flip = false;
+	if (val < 0) {
+		val = -val - 1;
+		flip = true;
+	}
+
+	let r = new VecU8(bytes);
+	for (var o = 0; o < bytes; ++o) {
+		r[o] = val % 256;
+		if (flip) {
+			r[o] = ~r[o] & 0xff;
+		}
+		val /= 256;
+	}
+	return r;
+}
 
 class VecU8 extends Uint8Array {
 	toJSON() {
@@ -110,4 +127,4 @@ function reviver(key, bland) {
 }
 
 module.exports = { VecU8, AccountId, Hash, VoteThreshold, SlashPreference, Moment, Balance,
-	BlockNumber, AccountIndex, Tuple, TransactionEra, Perbill, Permill, reviver }
+	BlockNumber, AccountIndex, Tuple, TransactionEra, Perbill, Permill, reviver, toLE }
